requirements.txt
llama-index
llama-index-vector-stores-qdrant
llama-index-llms-ollama
llama-index-llms-openai
llama-index-llms-anthropic
llama-index-embeddings-huggingface
llama-index-embeddings-ollama
python-dotenv
